Open-Address Hashing with Linear Probing and Double Hashing
Goals: In this exercise, there are multiple goals: 
(i) You will review open-address hashing and how to resolve collisions through linear probing and double-
hashing.  [No coding in this part; just review Section 12.2 of the textbook (pages 598-615) OR sections 5.1-5.5 
(with a special focus on 5.3 and 5.5), and the Powerpoint Hash Tables.pptx in the module for Week 12.] 
(ii) You are given a table class that initially implements linear probing.  You will modify a given table class 
to track the number of comparisons needed to find an index for a target (whether to insert, find, or delete) and 
report certain statistics.  This is Part a. 
(iii) You will make a very minor modification to the code to switch the collision resolution technique from 
linear probing to double-hashing and run it again to see how the performance now (in terms of number of 
comparisons) compares with that of linear probing in step ii.  This is Part b. 
(iv) You will be impressed with the performance difference!